LayoutTests/fast/canvas/html-in-canvas/drawElementImage-4arg-edge-cases.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>drawElementImage 4-arg — non-finite/zero/negative geometry is a silent no-op</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<canvas id="cvs" width="100" height="100" layoutsubtree>
  <div id="child" style="width:50px;height:50px;background:red"></div>
</canvas>
<script>
const cvs = document.getElementById("cvs");
const child = document.getElementById("child");
const ctx = cvs.getContext("2d");

function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

function assertNoOpReturnsIdentity(label, dx, dy, dw, dh) {
    let m;
    try {
        m = ctx.drawElementImage(child, dx, dy, dw, dh);
    } catch (e) {
        assert_unreached(`drawElementImage(child, ${dx}, ${dy}, ${dw}, ${dh}) threw ${e}`);
    }
    assert_true(m instanceof DOMMatrix, `${label}: returns DOMMatrix`);
    assert_equals(m.is2D, true, `${label}: is2D`);
    assert_true(m.isIdentity, `${label}: isIdentity`);
}

promise_test(async t => {
    await nextPaint();
    // Sentinel fill so we can verify pixels are untouched after no-op draws.
    ctx.fillStyle = "blue";
    ctx.fillRect(0, 0, 100, 100);

    assertNoOpReturnsIdentity("zero dwidth",      0, 0, 0,  50);
    assertNoOpReturnsIdentity("zero dheight",     0, 0, 50, 0);
    assertNoOpReturnsIdentity("negative dwidth",  0, 0, -10, 50);
    assertNoOpReturnsIdentity("negative dheight", 0, 0, 50, -10);
    // Sub-epsilon positive values: Chrome clamps these to zero in gfx::SizeF, so we
    // replicate the clamp (8 * FLT_EPSILON ≈ 9.5e-7). 1e-10 is well below the threshold.
    assertNoOpReturnsIdentity("sub-epsilon dwidth",  0, 0, 1e-10, 50);
    assertNoOpReturnsIdentity("sub-epsilon dheight", 0, 0, 50, 1e-10);
    assertNoOpReturnsIdentity("NaN dwidth",       0, 0, NaN, 50);
    assertNoOpReturnsIdentity("NaN dheight",      0, 0, 50, NaN);
    assertNoOpReturnsIdentity("Infinity dwidth",  0, 0, Infinity, 50);
    assertNoOpReturnsIdentity("Infinity dheight", 0, 0, 50, Infinity);
    assertNoOpReturnsIdentity("-Infinity dwidth", 0, 0, -Infinity, 50);
    assertNoOpReturnsIdentity("NaN dx",           NaN, 0, 50, 50);
    assertNoOpReturnsIdentity("NaN dy",           0, NaN, 50, 50);
    assertNoOpReturnsIdentity("Infinity dx",      Infinity, 0, 50, 50);

    const px = ctx.getImageData(50, 50, 1, 1).data;
    assert_array_equals([px[0], px[1], px[2], px[3]], [0, 0, 255, 255],
        "pixels unchanged by no-op draws (sentinel blue still present)");
}, "4-arg drawElementImage with non-finite/zero/negative geometry is a silent no-op returning identity");

promise_test(async t => {
    // Eligibility (canvas type, layoutsubtree, parent, snapshot) MUST run BEFORE geometry
    // validation, so that a non-HTMLCanvasElement context throws InvalidStateError instead
    // of silently returning identity for non-finite geometry.
    const offscreen = new OffscreenCanvas(100, 100);
    const octx = offscreen.getContext("2d");
    assert_throws_dom("InvalidStateError", () => {
        octx.drawElementImage(child, 0, 0, NaN, 50);
    }, "OffscreenCanvas + NaN dwidth still throws InvalidStateError (eligibility before geometry)");
}, "4-arg drawElementImage runs eligibility checks before geometry validation");
</script>
</body>
